requirements.txt
blinker==1.9.0
click==8.3.1
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
python-dateutil==2.9.0.post0
setuptools==80.10.2
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
wheel==0.46.3
pdfplumber==0.11.4
pytz
flask-compress==1.14
Flask-Caching==2.1.0
Pillow==10.4.0
Flask-Limiter==3.5.0
psycopg2-binary
gunicorn